Add Vectorize pass
Vectorize pass vectorizes consecutive loads or stores. This pass should
come after LoopUnrollPass and SimplifyCFGPass (InstCombine must not exist
here) to vectorize instructions inside loops. InstCombine pass should
follow to remove not used instructions (which were used to calculate
index).
This pass sinks vectorized instruction to after the last instruction
vectorized. Because of that, it doesn't vectorize load instructions if
base load instruction's use is before other load instructions. To
vectorize all possible loads, the uses of loads should be moved to
backward as much as possible.
Because of the problem written above, this pass doesn't vectorize most
loads for now, and need load move feature or pass, which will be
implemented in sprint 3.